wip: refactor code into modules
