+++ src/models/mod.rs
@@ -1,0 +1,8 @@
+pub mod goal;
+pub mod input;
+pub mod output;
+pub mod repetition;
+pub mod slot;
+pub mod slot_iterator;
+pub mod task;
+pub mod task_budgets;
