+++ src/models/slot.rs
@@ -11,6 +11,12 @@ pub struct Slot {
     pub end: NaiveDateTime,
 }
 
+#[derive(PartialEq, Eq, Clone)]
+struct SlotConflict {
+    slot: Slot,
+    num_conflicts: usize,
+}
+
 impl Sub for Slot {
     type Output = Vec<Slot>;
