+++ src/models/slot_iterator.rs
